Make it simpler to swap back to nvim-cmp if needed
